[Chore] Unfortunatelly, Linux support is canceled.
I have made the mapper_service implementation for windows really fast based on the sole fact that keys are u8. Linux uses u32 for keys, so after I've ported the program, it ate my memory and crashed. If somebody out there want's to go through the trouble of re-writting the whole mapper for linux (and windows) from scratch, feel free to open a PR. However, I currentlly don't have time for that, so Linux support is canceled for time beeing. Sorry :(